Removing all the unnecessary objects from fruit jelly scenario to focus on things that are needed. For now created a new file to store this scenario.
Also, I am following a new convension (which is better at least to me) where the scene / scenario is defined in a common demo script which both the sim and real scripts simply call into by passing in the right value for corresponding variable. This way, we don't need to maitain same copies of scenarios between sim and real and also we will be forced to make sim close/match to real.  We can discuss more the pros and cons of previous approach and this approach if anyones wishes.